move email validation to its own file
note that i moved the email-specific regex constants out of
notifications_utils/__init__.py and into the email_address file itself.
these are only used by the email_address code itself and in one place
in document download, and now that they're in a smaller file they seem
reasonable to place here
